contrib/abc-shell: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@43054 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/gnu/abc-shell/environ.c
@@ -48,7 +48,7 @@ copy_env(struct Hook *hook, APTR userdata, struct ScanVarsMsg *message)
 
                 ++env_size;
 
-                snprintf(buffer,size-1,"%s=%.*s", message->sv_Name, message->sv_VarLen, message->sv_Var);
+                snprintf(buffer,size-1,"%s=%.*s", message->sv_Name, (int)message->sv_VarLen, message->sv_Var);
 
                 *env  = buffer;
                 env++;
